修复详情页章节过多，页面向下滚动时崩溃的问题；修复无人评分的条目加载出现异常的问题

+++ Bangumi.Api/Models/Ep.cs
@@ -19,7 +19,7 @@ namespace Bangumi.Api.Models
         public int Type { get; set; }
 
         [JsonProperty("sort")]
-        public float Sort { get; set; }
+        public double Sort { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
